+++ web/usuaris/src/components/molecules/index.ts
@@ -1,4 +1,6 @@
 export { DateRange, type TDateRange } from './DateRange'
 export { ItineraryDropdown } from './ItineraryDropdown'
 export { StatusDropdown } from './StatusDropdown'
+export { RoleFilter } from './RoleFilter'
+
 export { Table, type TTable } from './Table'